requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2
python-decouple==3.8
drf-spectacular==0.26.5 